(GH-1060) Add shellwords for Windows MSVCRT
Shellwords was used to split a string into an array of strings, in preparation
to executing via Open3.  However Shellwords is Linux specific and does not work
on Windows reliably.  This commit adds a Shellwords utility based on a fork by
Lars Kanis. This commit also includes tests for the shellworlds utility.